Move his analysis to component
