+++ src/components/his/functions/HisTimeResampling.vue
@@ -1,0 +1,10 @@
+<template>
+  <v-card flat>
+    <v-card-title>Time Resampling</v-card-title>
+    <v-card-text>
+      <p>Time resampling is not yet implemented.</p>
+    </v-card-text>
+  </v-card>
+</template>
+
+<script setup lang="ts"></script>
